% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aae.data}
\alias{aae.data}
\title{aae.data: download and summarise streamflow data in R}
\description{
A tool to access streamflow data and associated data from
  the Victorian Water Measurement Information System, with customisable
  functions to calculate summary metrics at any resolution.

  See \code{\link{fetch_data}} and \code{\link{calculate}} for detailed
  examples.
}
