% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\alias{survey}
\alias{weekly}
\alias{monthly}
\alias{annual}
\alias{baseline}
\alias{days_below}
\alias{days_above}
\alias{rolling_range}
\title{Calculate custom metrics from streamflow data}
\usage{
calculate(value, date, resolution, fun = median, rescale = NULL, ...)

survey(season = 1:12, lag = 0, subset = NULL)

weekly(lag = 0, subset = NULL)

monthly(lag = 0, subset = NULL)

annual(season = 1:12, lag = 0, subset = NULL)

baseline(season = 1:12, subset = NULL)

days_below(x, threshold, ...)

days_above(x, threshold, ...)

rolling_range(x, lag, type = "ratio", ...)
}
\arguments{
\item{value}{a time series of streamflow data. \code{calculate} is written
and designed for daily data but may work in other cases}

\item{date}{a vector of dates in any format accepted by \pkg{lubridate}.
There should be one date for each observation in \code{value}}

\item{resolution}{a function defining the temporal resolution of calculations.
See details}

\item{fun}{a function (unquoted) used to calculate the final metric. Defaults
to \code{median} but any R function that returns a single numeric value will
work. The \pkg{aae.data} provides three alternatives: \code{days_below},
\code{days_above}, and \code{rolling_range}}

\item{rescale}{a \code{list} containing any of four elements that define how
a metric should be rescaled following calculation. Defaults to NULL, in which
case the output metric is not rescaled. The four options are: \code{subset},
\code{fun}, \code{season}, and \code{args}. See details for further information.}

\item{\dots}{any additional arguments to be passed to \code{fun}}

\item{season}{define months of year in which to calculate the metric (see details)}

\item{lag}{lag in calculation. Can be defined in any time period using the methods
in \pkg{lubridate} (see details)}

\item{subset}{a numeric vector defining years in which the metric should be calculated.
Can be a continuous or discontinuous set of years}

\item{x}{time series to be tested against threshold in \code{days_below} and
\code{days_above}, or a time series on which to calculate rolling ranges with
the \code{rolling_range} function}

\item{threshold}{defines level at which observations are considered
low or high}

\item{type}{specify how the difference (range) is calculated in
\code{rolling_range}. Defaults to "ratio" but setting any other
string will switch to absolute difference}
}
\description{
Calculate metrics at different temporal resolutions and with
  different lags from streamflow or related data downloaded with
  \code{fetch_data}.
}
\details{
\code{calculate} is the main function used to calculate metrics from
  input time series data. This function is designed to be highly flexible, and is
  used in conjunction with a set of \code{resolution} functions that define the
  temporal resolution and any subsetting of data within seasons or years. Currently
  available \code{resolution} functions are \code{survey}, \code{weekly}, \code{monthly},
  \code{annual}, and \code{baseline}. These functions calculate metrics by survey season,
  week, month, year, and averaged over all years, respectively. See below for further
  details of seasonal subsetting in the \code{survey} and \code{baseline} functions.

  The \code{rescale} argument gives the option of rescaling metrics by some other value
  calculated from the data, such as a long-term average. This is included to allow
  standardisation of metrics among rivers. \code{rescale} is a list that takes any of
  four elements. \code{subset} is a numeric vector that defines the years over which a
  long-term value should be calculated (defaults to all years included in \code{value}).
  \code{fun} defines the function used to calculate a long-term value (defaults to
  \code{median}). \code{season} is a numeric vector that defines the months in which
  the long-term value should be calculated (defaults to all months, \code{1:12}). \code{args}
  is a list containing any further arguments to pass to the \code{fun} defined in \code{rescale}
  (e.g. \code{na.rm = TRUE}).

The \code{survey} function defines seasons relative to survey times, so
  months 1-12 are Jan-Dec in the year prior to surveys (a year earlier than requested
  in \code{calculate}), and months 13-24 are Jan-Dec in the year of surveys.

  The \code{baseline} function defines seasons relative to the year of flow measurement,
  so months 1-12 are Jan-Dec in the year requested.

  The \code{lag} argument defaults to \code{years} for the \code{survey}, \code{monthly},
  \code{annual}, and \code{baseline} functions, and defaults to \code{weeks} for the
  \code{weekly} function. The \code{lag} argument is also used in \code{rolling_range},
  where it defaults to the unit of measurement in the input data (days, in most cases).

The \code{days_below} and \code{days_above} functions
  calculate the number of days below a defined threshold. Designed
  for use within \code{calculate}.

\code{rolling_range} calculates the maximum ratio or
  absolute difference in a variable over a specified lag.
}
\examples{
\dontrun{
# calculate basic metric (annual)

# calculate weekly metric

# calculate weekly metric with lag

# calculate survey metric

# example of num_below and num_above

# example of rolling range

# example of multiple metrics with dplyr

}
}
