% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_hydro.R
\name{fetch_hydro}
\alias{fetch_hydro}
\alias{fetch_rating}
\alias{check_quality}
\title{Download data from Victorian WMIS}
\usage{
fetch_hydro(
  sites,
  start,
  end,
  variables,
  include_missing = FALSE,
  include_qc = FALSE,
  options = list(),
  data_source = "A",
  state = "vic"
)

fetch_rating(sites, interval = 0.5, options = list(), state = "vic")

check_quality(data)
}
\arguments{
\item{sites}{a character or numeric vector of site codes (gauge numbers).
Only a single value can be used with \code{fetch_rating}}

\item{start}{start date for downloaded data. Date format is flexible but
for most reliable results use yyyy-mm-dd}

\item{end}{end date for downloaded data. Date format is flexible but
for most reliable results use yyyy-mm-dd}

\item{variables}{a character or numeric vector of variable names or codes.
Many variations are accounted for (e.g. streamflow, flow, discharge).
WMIS variable codes will work.}

\item{include_missing}{a \code{logical} defining whether to pad output data
to include dates where data are missing}

\item{include_qc}{a \code{logical} defining whether to return information on
quality control values. This information can be retrieved with
\link{check_quality}}

\item{options}{a \code{list} specifying one of four advanced options (only
varfrom and varto can be specified with \code{fetch_rating}):
- the interval type (defaults to "daily") for calculation of downloaded
  values
- the data_type (defaults to "mean") for aggregation of downloaded values
- the multiplier (defaults to "1") used to expand or shrink the requested
  interval
- varfrom and varto (default to NULL) to specify explicit variable
  conversions with WMIS variable codes. Both varfrom and varto must
  be provided}

\item{data_source}{database from which data are downloaded. Defaults to "A",
which covers most cases. The \code{list_datasources} function will list
all available data_sources for a given site}

\item{state}{state of interest; one of Victoria, New South Wales, or
Queensland}

\item{interval}{the gap between input levels used to generate the rating
table}

\item{data}{a \code{data.frame} returned by \code{fetch_hydro}}
}
\value{
a \code{data.frame} containing the downloaded flow data and relevant
  identifiers for each observation.

a \code{list} containing the downloaded gauge information.
}
\description{
Download data from the Victorian Water Measurement
  Information System (WMIS) for one or more variables at one or more sites.
}
\details{
Streamflow and related data (e.g. temperature) are recorded as
  instantaneous values, so must be aggregated prior to download. The default
  settings specify daily means, but other options are available with the
  \code{options} parameter.

  Dates and variable names are interpreted loosely. Dates will recognise any
  of the following formats: \code{"ymd"}, \code{"dmy"}, \code{"y"},
  \code{"ymd_HMS"}, and \code{"ymd_HM"}. Variables will ignore case and
  remove all spaces and underscores. Once tidied, any variable name
  containing any of the following will be recognised: \code{"flow"},
  \code{"discharge"}, \code{"depth"}, \code{"height"}, \code{"temp"},
  \code{"cond"}, \code{"ox"}, \code{"do"}, and \code{"turb"}.

Rating tables are used to convert gauge heights to discharge.
  These tables would rarely need to be used by the end user but may provide
  useful information in some cases. In addition, this function returns
  additional information about a gauge, including the cease-to-flow (CTF)
  threshold, that is, the level (in m) at which the system will have zero
  flow.

tabulates data by QC code, with descriptions of each QC code.
}
\examples{
\dontrun{

# grab data for three variables at three sites
flow_data <- fetch_hydro(
  sites = c("405232", "406276", "406278"),
  start = "2004-01-01",
  end = "2020-04-21",
  variables = c("flow", "temp", "depth")
)

# try a different date range
flow_data <- fetch_hydro(
  sites = c("405232", "406276", "406278"),
  start = "1991-01-01",
  end = "1993-04-21",
  variables = c("flow", "temp", "depth")
)

# or a different date format
flow_data <- fetch_hydro(
  sites = c("405232", "406276", "406278"),
  start = "01-01-1994",
  end = "1996",
  variables = c("flow", "temp", "depth")
)

# repeat, but fill gaps with placeholders
flow_data <- fetch_hydro(
  sites = c("405232", "406276", "406278"),
  start = "2004-01-01",
  end = "2020-04-21",
  variables = c("flow", "temp", "depth"),
  include_missing = TRUE
)

# check for data from the telemetry data source
flow_data <- fetch_hydro(
  sites = c("405232", "406276", "406278"),
  start = "2004-01-01",
  end = "2020-04-21",
  variables = c("flow", "temp", "depth"),
  data_source = "TELEM"
)

# advanced options: return weekly median values
#   rather than daily mean values
flow_data <- fetch_hydro(
  sites = c("405232", "406276", "406278"),
  start = "2004-01-01",
  end = "2020-04-21",
  variables = c("flow", "temp", "depth"),
  options = list(
    interval = "weekly",
    data_type = "median"
  )
)

# advanced options: specify variables explicitly
#   by WMIS codes
flow_data <- fetch_hydro(
  sites = c("405232", "406276", "406278"),
  start = "2004-01-01",
  end = "2020-04-21",
  variables = c("flow", "temp", "depth"),
  options = list(
    varfrom = "100.00",
    varto = "141.00"
  )
)

}

\dontrun{

# fetch the rating table for several gauges, at an interval of
#   0.5 m
gauge_info <- fetch_rating(
  sites = 405232,
  interval = 0.5,
  options = list(
    varfrom = "100.00",
    varto = "141.00"
  )
)

}

\dontrun{

# download some flow data
flow_data <- fetch_hydro(
  sites = c("405232", "406276", "406278"),
  start = "1991-01-01",
  end = "1993-04-21",
  variables = c("flow", "temp", "depth")
)

# and print a table of QC codes to the console
check_quality(flow_data)

}
}
