% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\alias{impute_hydro}
\title{Impute missing streamflow or water temperature data}
\usage{
impute_hydro(x, y, method)
}
\description{
Functions to impute missing hydrological data using
  nearby weather data or rolling means.
}
