% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_database}
\alias{query_database}
\alias{list_variables}
\alias{list_datasources}
\title{Interact with WMIS database from R}
\usage{
list_variables(sites, data_source = NULL, state = "vic")

list_datasources(sites, state = "vic")
}
\arguments{
\item{sites}{a character or numeric vector of site codes (gauge numbers)}

\item{data_source}{database from which data are downloaded. Defaults to
NULL, which will check all available data sources. The
\code{list_datasources} function lists all available data_sources for
a given site}

\item{state}{state of interest; one of Victoria, New South Wales, or
Queensland}
}
\value{
a \code{data.frame} containing information on variables and
  data sources at each site.

a \code{list} containing information on data sources at each site.
}
\description{
Functions to prepare and submit queries to the Victorian
  Water Measurement Information System (WMIS) database.
}
\details{
\code{list_variables} queries the WMIS database to identify all
  variables avaiable at a given site.

\code{list_datasources} queries the WMIS database to identify all
  data sources avaiable at a given site.
}
\examples{
\dontrun{
  # list all variables at three sites
  list_variables(
    sites = c("405232", "406276", "406278"),
  )

  # list all variables at three sites from telemetry data
  list_variables(
    sites = c("405232", "406276", "406278"),
    data_source = "TELEM"
  )

  # list all data sources at three sites
  list_datasources(
    sites = c("405232", "406276", "406278"),
  )
}
}
