% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\alias{by_median}
\alias{by_mean}
\alias{by_max}
\alias{by_generic}
\title{Functions to define long-term averages to rescale metrics}
\usage{
by_median(subset, season = 1:12, ...)

by_mean(subset, season = 1:12, ...)

by_max(subset, season = 1:12, ...)

by_generic(subset, season = 1:12, fun = median, ...)
}
\arguments{
\item{subset}{a numeric vector defining years in which the metric
should be calculated. Can be a continuous or discontinuous set
of years}

\item{season}{define months of year in which to calculate the
metric (see details)}

\item{\dots}{additional arguments passed to \code{median},
\code{mean}, \code{max}, or \code{fun} (e.g. `na.rm = TRUE`)}

\item{fun}{a function (unquoted) used to calculate the final metric. Defaults
to \code{median} but any R function that returns a single numeric value
will work. The \pkg{aae.hydro} provides three additional options:
\code{days_below}, \code{days_above}, and \code{rolling_range}}
}
\value{
a list with four elements: subset, fun, season, and args
}
\description{
Functions to define long-term average values used to
  standardise metrics in \code{calculate}.
}
