% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\alias{generate_scenario}
\alias{resample_scenario}
\alias{water_year}
\alias{define_transition}
\title{Generate hydrological scenarios from data}
\usage{
generate_scenario(
  x,
  date,
  thresholds,
  timing,
  duration,
  change,
  n,
  context = NULL,
  context_data = NULL,
  resolution = water_year,
  ...
)

resample_scenario(
  x,
  date,
  n,
  transition,
  context,
  context_data = NULL,
  resolution = water_year,
  rep = 1,
  init = NULL,
  init_year = NULL,
  y = NULL,
  ...
)

water_year(x)

define_transition(
  x,
  date,
  context,
  context_data = NULL,
  resolution = water_year,
  ...
)
}
\arguments{
\item{x}{discharge data from gauge or a date vector in the case of
\code{water_year}}

\item{date}{formatted date corresponding to gauge flows}

\item{thresholds}{list of flow thresholds to be combined with
timing and context to modify gauge flows. See Details for
usage.}

\item{timing}{a list specifying the timing of each threshold;
currently restricted to (integer) calendar months (1 = Jan,
12 = Dec)}

\item{duration}{a list specifying the duration for which
each threshold is applied}

\item{change}{a list specifying the minimum or maximum daily change
allowed in a hydrological sequence}

\item{n}{number of time periods to sample}

\item{context}{a character vector denoting the context for each
time period or a function that can be applied to discharge data
(default) or context data (if provided) to determine the context
for each time period}

\item{context_data}{additional data used to define context
when this depends on factors other than discharge in the
defined time period (e.g. rainfall, temperature, or
antecedent discharge). Note that context data will be
compiled according to the specified resolution and context
should be provided as a character vector if this is not
sufficient}

\item{resolution}{a function to classify dates into discrete
units (e.g. calendar years, water years, months) over which
the context, timing, and thresholds can be applied. See
\code{water_year} for an example}

\item{\dots}{additional arguments passed to \code{context}}

\item{transition}{matrix with one row and column for each level
of context, defining the transition probabilities between each
context pair (e.g. dry to dry, dry to wet). Matrix structures
should reflect columns moving to rows (i.e. x_new = A x_old),
and matrices will be internally standardised so that each
column sums to one}

\item{rep}{number of replicate scenarios to generate}

\item{init}{string denoting the initial context}

\item{init_year}{initial year for resampled dates}

\item{y}{extra variable to be resampled according to transitions
and contexts defined by \code{x}}
}
\description{
Functions to update hydrological data to reflect
  management scenarios (esp. environmental watering) and resample
  from existing data to create hypothetical climate change scenarios
}
\details{
\code{generate_scenario} is a function to define
  flow scenarios from gauge data based on a series of rules
  establishing flow thresholds, timings, and durations.

  Each element of thresholds can be a named list or vector
  with three values. Vectors denote lower, upper, and maximum thresholds
  to modify discharge values, e.g., a value of c(1, 10, 100) will set
  all values below 1 to a random value between 1 and 10 and all values
  above 100 to 100. If only one of these bounds is required
  vectors can include -Inf and Inf or NA values, which are ignored.
  Named lists are used when context is necessary. In this case, the
  context argument assigns each time period to a category that
  determines which threshold is applied.

  Thresholds are applied first-to-last, so precedence is given to
  those later in the list.

  Timing must be provided for each threshold, and denotes the
  months in which each threshold is applied. Months are defined
  based on calendar months (1 to 12) and must be a full vector
  including all relevant months, not just the bounds. For example,
  May to August is specified by c(5:8) not c(5, 8).

  Duration (in days) of each threshold is used primarily to
  define large flow events (freshes and overbanks), which often
  occur for several days within a month or season. If threshold
  instead defines a longer flow duration (e.g. a baseflow), then
  duration can be set to Inf. The parameter n specifies the
  number of times a given event can occur in a given time
  period. This is a vector with two values that specify lower
  and upper bounds on the number of events, and can be set to NA
  in cases where it is not required (e.g. baseflows).

  Change can be used to specify an upper bound on the daily
  change in flow. If a daily change exceeds this value, the
  second day (i.e. the day of the change) is replaced with
  the mean of surrounding values (in an iteratively widening window)
  until the difference is less than that specified by change.

  Resolution is not checked internally for consistency with timing
  or context, so it is assumed that the provided arguments are
  appropriate and compatible.

\code{resample_scenario} is a function to resample from
  data (e.g. discharge or water temperature) to create sequences
  representing plausible future climates. Future climates are
  defined by transitions between different contexts (e.g. dry and
  wet years), with transition probabilities (and associated return
  times) allowing a range of scenarios (e.g. longer droughts,
  more frequent extreme events).

  Both the resolution and context can be changed to match specific
  applications, such as resampling months or weeks and specifying
  nuanced contexts based on any features of interest in hydrological
  data or other data (provided as \code{context_data}).

\code{water_year} is a function to define
  water years from date data. Included here to demonstrate
  how similar functions might be specified.

\code{define_transition} is a function to define
 transition probabilities from existing data and contexts.
}
\examples{
\dontrun{
#

}
\dontrun{
#

}
\dontrun{
#

}
\dontrun{
#

}
}
